requirements.txt
Django
django-cors-headers
djangorestframework
django-user-agents
dnspython
pymongo
python-crontab
python-dateutil
pytz
requests
six
tzdata
urllib3
vine
wcwidth
zipp
markdown
stringgenerator
django-webpack-loader
python-dotenv
django-apscheduler
apscheduler
gunicorn
pycryptodome
gunicorn
flask
pillow
numpy
passporteye